refactor: modularize into 8 files — config, db, error, fmt, input, report, watcher
Split monolithic main.rs (1533 lines) into focused modules:
- main.rs: CLI dispatch only (79 lines)
- config.rs: Config parsing, Category enum, TitleRule, classify()
- db.rs: SessionSnapshot, migrations, insert_event()
- error.rs: Error enum (thiserror)
- fmt.rs: Duration/percentage/bar formatting helpers
- input.rs: InputSnapshot, evdev monitoring, named constants
- report.rs: App struct, today/metrics/timeline/report commands
- watcher.rs: Niri IPC event loop, WindowInfo, backoff

Style: impl Display/FromStr for Category, struct over tuples,
Config::classify() method, App::open() for init boilerplate.
